refactor(Imports): Changing the core files imports to project folder core
