+++ app/db/base.py
@@ -1,7 +1,7 @@
 from sqlalchemy import create_engine
 from sqlalchemy.orm import Session
 
-from app.settings import Settings
+from app.core.settings import Settings
 
 engine = create_engine(Settings().DATABASE_URL)
